requirements.txt
openai>=1.0.0
rich>=13.0.0
datasets>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
anthropic>=0.18.0
google-generativeai>=0.3.0

